Dream11-Team-Predictor-main/src/UI/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.1.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.9.0
boto3==1.34.76
botocore==1.34.76
bs4==0.0.2
cached-property==1.5.2
cachetools==4.2.4
celery==5.3.6
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.4.4
cryptography==35.0.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==5.0.0
flower==2.0.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.0
fuzzywuzzy==0.18.0
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
google-ai-generativelanguage==0.6.4
google-api-core==1.34.1
google-api-python-client==2.142.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.19.0
google-cloud-core==1.7.3
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-cloud-translate==2.0.1
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.48.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
humanize==4.10.0
identify==2.6.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
instructor==1.6.2
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.3.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
kombu==5.3.6
langchain==0.3.4
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langsmith==0.1.136
linkedin-api==2.0.3
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
narwhals==1.13.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
notebook==7.2.2
notebook-as-pdf==0.5.0
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
openai==1.52.0
opencv-python==4.10.0.84
ordered-set==4.1.0
orjson==3.10.11
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandoc==2.4
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathos==0.3.2
patsy==0.5.6
pdf2image==1.17.0
peft==0.11.1
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.4.0
pip-review==1.3.0
pipreqs==0.4.13
platformdirs==3.11.0
plotly==5.20.0
pluggy==1.5.0
plumbum==1.9.0
ply==3.11
pmdarima==2.0.4
pooch==1.8.0
portkey-ai==1.7.2
pox==0.3.4
ppft==1.7.6.8
pre_commit==4.0.1
preshed==3.0.9
proglog==0.1.10
progress==1.6
prometheus-client==0.17.1
promise==2.3
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.6
pure-eval==0.2.2
py-dreambooth==0.2.8
pyarrow==18.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyee==11.1.1
pygame==2.6.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.1.0
pymongo==4.6.3
pynndescent==0.5.11
pyOpenSSL==22.0.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypinyin==0.49.0
pyppeteer==2.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.1.1
pytest-cov==5.0.0
python-crfsuite==0.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-espncricinfo==0.5.8
python-json-logger==2.0.7
python-ondc==0.0.1
pytz==2024.2
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.2
pyzmq==24.0.1
rapidfuzz==3.9.1
redis==5.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
sagemaker==2.223.0
schema==0.7.7
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.13.0
Send2Trash==1.8.2
sentry-sdk==2.5.1
service-identity==23.1.0
setproctitle==1.3.3
shapely==2.0.3
shellescape==3.8.1
silx==1.1.2
six==1.16.0
smart-open==6.4.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
sqlparse==0.4.3
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.27.2
SudachiDict-core==20230927
SudachiPy==0.6.7
swapper==1.3.0
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
sympy==1.12
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
thefuzz==0.22.1
thinc==8.2.1
threadpoolctl==3.3.0
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.0
toolz==0.12.0
torch==2.2.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.6
trainer==0.0.33
traitlets==5.14.3
transformers==4.41.2
trio==0.24.0
trio-websocket==0.11.1
TTS==0.21.3
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typer==0.9.0
types-python-dateutil==2.8.19.14
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.5
Unidecode==1.3.7
uri-template==1.3.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.23.2
validators==0.22.0
vine==5.1.0
virtualenv==20.27.0
waitress==3.0.0
wandb==0.17.2
wasabi==1.1.2
watchdog==5.0.3
wcwidth==0.2.8
weasel==0.3.4
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.22
xgboost==2.1.1
yarg==0.1.9
yarl==1.15.5
zerobouncesdk==1.1.0
zipp==3.20.2
zope.interface==6.0
